separated collection statistics group into its owen component
