+++ ui/src/components/Collection/CollectionMetadataPanel.scss
@@ -2,37 +2,25 @@
 @import "app/mixins.scss";
 
 .CollectionMetadataPanel {
-  &__vertical-container {
-    @media screen and (max-width: $aleph-screen-md-max-width) {
-      margin-right: 20px;
-      width: 50%;
-      height: 100%;
-
-    }
-    & > *:first-child {
-      @media screen and (max-width: $aleph-screen-md-max-width) {
-        margin-bottom: 20px;
-      }
-    }
+  @media screen and (max-width: $aleph-screen-sm-max-width) {
+    display: flex;
+    flex-flow: row wrap;
+    margin: -$aleph-content-padding/2;
   }
 
   &__item {
-    margin-bottom: 20px;
-    padding: 15px;
+    margin-bottom: $aleph-content-padding;
+    padding: $aleph-grid-size*1.5;
     background: $light-gray5;
     border-radius: 5px;
     flex: 1 1 auto;
 
-    &:first-child {
-      margin-left: 0 !important;
-    }
-
-    @media screen and (max-width: $aleph-screen-md-max-width) {
-      margin-bottom: 0;
+    @media screen and (max-width: $aleph-screen-sm-max-width) {
+      margin: $aleph-content-padding/2;
     }
 
     &.mobile-hide {
-      @media screen and (max-width: $aleph-screen-md-max-width) {
+      @media screen and (max-width: $aleph-screen-sm-max-width) {
         display: none;
       }
     }
@@ -45,9 +33,5 @@
       font-size: 14px !important;
       margin-bottom: 0 !important;
     }
-  }
-
-  &__actions {
-    margin-top: 20px;
   }
 }
