+++ ui/src/components/Collection/CollectionOverviewMode.scss
@@ -1,57 +1,39 @@
 @import "app/variables.scss";
+@import "app/mixins.scss";
+
+$metadata-width: $aleph-grid-size*30;
+$metadata-padding: $aleph-content-padding;
+$metadata-breakpoint: $aleph-screen-sm-max-width;
 
 .CollectionOverviewMode {
-  column-count: 4;
-  column-gap: 1.5em;
-  table-layout : fixed;
-  column-fill: balance;
-  height: 100%;
-
-  @media screen and (max-width: $aleph-screen-lg-max-width) {
-    column-count: 3;
+  display: flex;
+  flex-flow: row;
+  width: 100%;
+  @media screen and (max-width: $metadata-breakpoint) {
+    flex-direction: column-reverse;
   }
 
-  @media screen and (max-width: $aleph-screen-md-max-width) {
-    column-count: 2;
-  }
-
-  @media screen and (max-width: $aleph-screen-sm-max-width) {
-    column-count: 1;
-  }
-
-  &__item {
-    -webkit-column-break-inside: avoid;
-    page-break-inside: avoid;
-    break-inside: avoid;
-
-    padding-bottom: 1.5em;
-    padding-top: 1px;
-
-    &__text-content {
-      background-color: $aleph-breadcrumbs-background !important;
-      padding: $aleph-grid-size*2;
-      border-radius: $aleph-border-radius;
-
-      &__divider {
-        height: 1px;
-        width: 100%;
-        border-bottom: 1px solid $aleph-border-color;
-        margin: 15px 0;
-
-        &:last-child {
-          display: none;
-        }
-      }
-    }
-
-    & > * {
-      // Firefox fix to avoid break inside column
-      display: inline-block;
-      width: 100%;
+  &__main, &__secondary {
+    @media screen and (max-width: $metadata-breakpoint) {
+      width: 100% !important;
     }
   }
 
-  &__summary {
-    margin-bottom: $aleph-grid-size*2;
+  &__main {
+    width: calc(100% - #{$metadata-width} - #{$metadata-padding});
+  }
+
+  &__secondary {
+    width: $metadata-width;
+    @include rtlSupportInvertedProp(margin, left, $metadata-padding, 0);
+    @include rtlSupportInvertedProp(padding, left, $metadata-padding, 0);
+    @include rtlSupportInvertedProp(border, left, 1px solid $aleph-border-color, none);
+
+    @media screen and (max-width: $metadata-breakpoint) {
+      padding: 0 0 $metadata-padding !important;
+      margin: 0 0 $metadata-padding !important;
+      border: none !important;
+      border-bottom: 1px solid $aleph-border-color !important;
+    }
   }
 }
