+++ ui/src/components/Collection/CollectionStatisticsGroup.scss
@@ -1,0 +1,53 @@
+@import "app/variables.scss";
+
+.CollectionStatisticsGroup {
+  column-count: 3;
+  column-gap: 1.5em;
+  table-layout : fixed;
+  column-fill: balance;
+  height: 100%;
+
+  @media screen and (max-width: $aleph-screen-lg-max-width) {
+    column-count: 2;
+  }
+
+  @media screen and (max-width: $aleph-screen-md-max-width) {
+    column-count: 1;
+  }
+
+  &__item {
+    -webkit-column-break-inside: avoid;
+    page-break-inside: avoid;
+    break-inside: avoid;
+
+    padding-bottom: 1.5em;
+    padding-top: 1px;
+
+    &__text-content {
+      background-color: $aleph-breadcrumbs-background !important;
+      padding: $aleph-grid-size*2;
+      border-radius: $aleph-border-radius;
+
+      &__divider {
+        height: 1px;
+        width: 100%;
+        border-bottom: 1px solid $aleph-border-color;
+        margin: 15px 0;
+
+        &:last-child {
+          display: none;
+        }
+      }
+    }
+
+    & > * {
+      // Firefox fix to avoid break inside column
+      display: inline-block;
+      width: 100%;
+    }
+  }
+
+  &__summary {
+    margin-bottom: $aleph-grid-size*2;
+  }
+}
