+++ ui/src/components/Collection/CollectionViews.jsx
@@ -13,6 +13,7 @@ import CollectionMappingsMode from 'components/Collection/CollectionMappingsMode
 import CollectionEntitiesMode from 'components/Collection/CollectionEntitiesMode';
 import CollectionXrefMode from 'components/Collection/CollectionXrefMode';
 import CollectionEntitySetsIndexMode from 'components/Collection/CollectionEntitySetsIndexMode';
+import CollectionStatisticsGroup from 'components/Collection/CollectionStatisticsGroup';
 import FacetedEntitySearch from 'components/EntitySearch/FacetedEntitySearch';
 import collectionViewIds from 'components/Collection/collectionViewIds';
 import { queryCollectionEntities, collectionXrefFacetsQuery } from 'queries';
@@ -69,10 +70,11 @@ class CollectionViews extends React.Component {
               <Icon icon="grouped-bar-chart" className="left-icon" />
               <FormattedMessage id="entity.info.overview" defaultMessage="Overview" />
             </>}
-          panel={isCasefile
-            ? <InvestigationOverviewMode collection={collection} />
-            : <CollectionOverviewMode collection={collection} />
-          }
+          panel={(
+            <CollectionOverviewMode collection={collection}>
+              {!isCasefile && <CollectionStatisticsGroup collection={collection} />}
+            </CollectionOverviewMode>
+          )}
         />
         {showDocumentsTab && (
           <Tab
